requirements.txt
pyyaml==6.0
faker==8.12.1
tqdm==4.62.3
nltk==3.6.3